+++ CHANGELOG.md
@@ -19,8 +19,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - `remove-request-query-parameter` scriptlet to remove query parameters
   from `fetch` and `XMLHttpRequest` requests [#329].
 
+### Fixed
+
+- Anti-adblock detection in `spoof-css` scriptlet
+  by using cloaked bound functions instead of Proxies [#422].
+
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.2.15...HEAD
 [#329]: https://github.com/AdguardTeam/Scriptlets/issues/329
+[#422]: https://github.com/AdguardTeam/Scriptlets/issues/422
 [#461]: https://github.com/AdguardTeam/Scriptlets/issues/461
 
 ## [v2.2.15] - 2026-01-22